Being hardcore

+++ steal_limelight.sh
@@ -1,30 +1,42 @@
 #!/bin/sh
 
-# 
-# This essentially says:
-# "Get me everything in and next to https://www.gov.uk/performance/services"
-# "... but IGNORE the Dashboard and Transactions Explorer"
-# 
-# What you should end up with from the wget is a set of .html files which link locally to each other
-# Most or all of css / js / images etc will come from https://gov.uk, so you will STILL NEED a web connection
-# 
+echo "ATTENTION! This script will DELETE your 'performance' folder if it is present!"
+echo "Make sure that you MOVE the folder elsewhere if you want to keep any prototyped work!"
+echo
+read -p "Are you ready?" -n 1 -r
 
-wget --mirror --page-requisites --adjust-extension --no-parent --convert-links --exclude-directories=/performance/transactions-explorer/,/performance/dashboard/ --reject=dashboard,transactions-explorer https://www.gov.uk/performance/services
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    # do dangerous stuff - delete the performance folder:
 
-echo "Limelight has been grabbed!"
+    rm -R performance
 
-# 
-# Next up, we're going to move the /performance/ folder into the root
-# 
+    # 
+    # This essentially says:
+    # "Get me everything in and next to https://www.gov.uk/performance/services"
+    # "... but IGNORE the Dashboard and Transactions Explorer"
+    # 
+    # What you should end up with from the wget is a set of .html files which link locally to each other
+    # Most or all of css / js / images etc will come from https://gov.uk, so you will STILL NEED a web connection
+    # 
 
-echo "Moving files..."
-mv www.gov.uk/performance/ .
-echo "Done."
+    wget --mirror --page-requisites --adjust-extension --no-parent --convert-links --exclude-directories=/performance/transactions-explorer/,/performance/dashboard/ --reject=dashboard,transactions-explorer https://www.gov.uk/performance/services
 
-# 
-# Then delete the leftover www.gov.uk folder (and robots.txt)
-# 
+    echo "Limelight has been grabbed!"
 
-echo "Deleting unused files..."
-rm -R www.gov.uk
-echo "Done."
+    # 
+    # Next up, we're going to move the /performance/ folder into the root
+    # 
+
+    echo "Moving files..."
+    mv www.gov.uk/performance/ .
+    echo "Done."
+
+    # 
+    # Then delete the leftover www.gov.uk folder (and robots.txt)
+    # 
+
+    echo "Deleting unused files..."
+    rm -R www.gov.uk
+    echo "Done."
+fi
